second version of pool allocator
git-svn-id: http://svn.berlios.de/svnroot/repos/bvr20983/trunk@203 01137330-e44e-0410-aa50-acf51430b3d2

+++ src/util/yanew.cpp
@@ -25,7 +25,7 @@
 /**
  *
  */
-void* operator new(size_t bytes,bvr20983::util::YAAllocatorBase* a,LPCTSTR filename, int lineno)
+void* operator new(size_t bytes,bvr20983::util::YAAllocatorBase* a,LPCTSTR filename,unsigned short lineno)
 { bvr20983::util::YAAllocatorResult* result;
 
   if( NULL!=a )
@@ -39,7 +39,7 @@ void* operator new(size_t bytes,bvr20983::util::YAAllocatorBase* a,LPCTSTR filen
 /**
  * is only called, if exception is thrown in constructor
  */
-void operator delete(void* p,bvr20983::util::YAAllocatorBase* a,LPCTSTR filename, int lineno)
+void operator delete(void* p,bvr20983::util::YAAllocatorBase* a,LPCTSTR filename,unsigned short lineno)
 { //OutputDebugFmt(_T("%s[%d] delete(0x%lx)\n"),filename,lineno,p);
 
   if( NULL!=a )
